+++ NoVacancy/Controllers/CarritoLineaController.cs
@@ -4,6 +4,17 @@ using NoVacancy.Data;
 using Microsoft.EntityFrameworkCore;
 using System.Linq;
 
+/*
+ * 
+ * NOTA IMPORTANTÍSIMA:
+ * C.Id (con la I mayúscula) hace referencia
+ * al ID del Usuario, Identity la maneja así,
+ * y además es un String, no un int
+ * 
+ * Maldigo a Identity
+ * 
+ */
+
 namespace NoVacancy.Controllers
 {
     public class CarritoLineaController : Controller
@@ -23,12 +34,11 @@ namespace NoVacancy.Controllers
         // GET: CarritoLinea/ShoppingCart
         public async Task<IActionResult> ShoppingCart()
         {
-            // Suponiendo que el id del usuario está en la sesión
-            int? usuarioId = HttpContext.Session.GetInt32("UrusarioId");
+            string? usuarioId = HttpContext.Session.GetString("Id");
             if (usuarioId == null)
-                return RedirectToAction("Login", "Usurario");
+                return RedirectToAction("Login", "Usuario");
 
-            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.idUsuario == usuarioId);
+            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.Id == usuarioId);
             if (carrito == null)
                 return View(new List<CarritoLinea>());
 
@@ -44,14 +54,18 @@ namespace NoVacancy.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(int idProducto, int cantidad = 1)
         {
-            int usuarioId = 1;
-            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.idUsuario == usuarioId);
+            string? usuarioId = HttpContext.Session.GetString("Id");
+            if (usuarioId == null)
+                return RedirectToAction("Login", "Usuario");
+
+            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.Id == usuarioId);
             if (carrito == null)
             {
-                carrito = new CarritoCabecera { idUsuario = usuarioId };
+                carrito = new CarritoCabecera { Id = usuarioId };
                 _context.CarritosCabecera.Add(carrito);
                 await _context.SaveChangesAsync();
             }
+
             var linea = await _context.CarritosLineas.FirstOrDefaultAsync(l => l.idCarrito == carrito.idCarrito && l.idProducto == idProducto);
             if (linea != null)
             {
@@ -63,6 +77,7 @@ namespace NoVacancy.Controllers
                 linea = new CarritoLinea { idCarrito = carrito.idCarrito, idProducto = idProducto, cantidad = cantidad };
                 _context.CarritosLineas.Add(linea);
             }
+
             await _context.SaveChangesAsync();
             return RedirectToAction("Shopping_cart");
         }
@@ -71,10 +86,14 @@ namespace NoVacancy.Controllers
         [HttpPost]
         public async Task<IActionResult> Update(int idProducto, int cantidad)
         {
-            int usuarioId = 1;
-            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.idUsuario == usuarioId);
+            string? usuarioId = HttpContext.Session.GetString("Id");
+            if (usuarioId == null)
+                return RedirectToAction("Login", "Usuario");
+
+            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.Id == usuarioId);
             if (carrito == null)
                 return RedirectToAction("Shopping_cart");
+
             var linea = await _context.CarritosLineas.FirstOrDefaultAsync(l => l.idCarrito == carrito.idCarrito && l.idProducto == idProducto);
             if (linea != null)
             {
@@ -82,6 +101,7 @@ namespace NoVacancy.Controllers
                 _context.CarritosLineas.Update(linea);
                 await _context.SaveChangesAsync();
             }
+
             return RedirectToAction("Shopping_cart");
         }
 
@@ -89,16 +109,21 @@ namespace NoVacancy.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(int idProducto)
         {
-            int usuarioId = 1;
-            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.idUsuario == usuarioId);
+            string? usuarioId = HttpContext.Session.GetString("Id");
+            if (usuarioId == null)
+                return RedirectToAction("Login", "Usuario");
+
+            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.Id == usuarioId);
             if (carrito == null)
                 return RedirectToAction("Shopping_cart");
+
             var linea = await _context.CarritosLineas.FirstOrDefaultAsync(l => l.idCarrito == carrito.idCarrito && l.idProducto == idProducto);
             if (linea != null)
             {
                 _context.CarritosLineas.Remove(linea);
                 await _context.SaveChangesAsync();
             }
+
             return RedirectToAction("Shopping_cart");
         }
 
@@ -106,14 +131,18 @@ namespace NoVacancy.Controllers
         [HttpPost]
         public async Task<IActionResult> Clear()
         {
-            int usuarioId = 1;
-            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.idUsuario == usuarioId);
+            string? usuarioId = HttpContext.Session.GetString("Id");
+            if (usuarioId == null)
+                return RedirectToAction("Login", "Usuario");
+
+            var carrito = await _context.CarritosCabecera.FirstOrDefaultAsync(c => c.Id == usuarioId);
             if (carrito != null)
             {
                 var lineas = _context.CarritosLineas.Where(l => l.idCarrito == carrito.idCarrito);
                 _context.CarritosLineas.RemoveRange(lineas);
                 await _context.SaveChangesAsync();
             }
+
             return RedirectToAction("Shopping_cart");
         }
     }
